# Conflicts:
#	src/app/app-routing.module.ts
#	src/app/app.component.html
#	src/app/app.module.ts
#	src/app/models/cabinet.spec.ts
#	src/app/models/composite.spec.ts
#	src/app/models/doctor-specialty.spec.ts
#	src/app/models/doctor.spec.ts
#	src/app/models/patient.spec.ts
#	src/app/models/schedule.spec.ts
#	src/app/models/specialty.spec.ts
#	src/app/models/timeslot.spec.ts
#	src/app/models/user.spec.ts
#	src/app/services/user-service.service.ts

+++ src/app/lists/cabinet-list/cabinet-list.component.ts
@@ -12,10 +12,10 @@ export class CabinetListComponent implements OnInit {
 
   displayedColumns: string[] = ['cabinet', 'delete'];
   dataSource;
-
-  constructor(private route: ActivatedRoute, private router: Router, private cabinetService: CabinetService) {}
-
   @ViewChild(MatSort, {static: true}) sort: MatSort;
+
+  constructor(private route: ActivatedRoute, private router: Router, private cabinetService: CabinetService) {
+  }
 
   ngOnInit() {
     this.reloadData();
@@ -29,13 +29,14 @@ export class CabinetListComponent implements OnInit {
   }
 
   deleteCabinet(id: number) {
-    this.cabinetService.deleteCabinet(id)
+    if (confirm('Уверены что хотите удалить?')) {
+      this.cabinetService.deleteCabinet(id)
       .subscribe(
-        data => {
-          console.log(data);
-          this.reloadData();
-        },
-        error => console.log(error));
+          data => {
+            console.log(data);
+            this.reloadData();
+          },
+          error => console.log(error));
+    }
   }
-
 }
